# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Marketing Simulation Lab API",
    "description": "API documentation for Digital Marketing Simulation Lab backend",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/users/profile": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Update current user's profile display name and personal details",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "email": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "universityRole": {
                    "type": "string"
                  },
                  "age": {
                    "type": "integer"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Profile updated successfully",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string",
                          "nullable": true
                        },
                        "universityRole": {
                          "type": "string",
                          "nullable": true
                        },
                        "age": {
                          "type": "integer",
                          "nullable": true
                        },
                        "gender": {
                          "type": "string",
                          "nullable": true
                        },
                        "category": {
                          "type": "string",
                          "nullable": true
                        },
                        "institution": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/join-class": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Associates a student to a class room via class invite code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "inviteCode"
                ],
                "properties": {
                  "inviteCode": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully joined class",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Successfully joined class",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "classId": {
                      "type": "string"
                    },
                    "className": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/assign-role": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Allows administrators to reassign user system roles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "role"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Role assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Role assigned successfully",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Lists all users in the system (Admin only)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Users list",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "joinedAt": {
                            "type": "string"
                          },
                          "lastLogin": {
                            "type": "string"
                          },
                          "classCount": {
                            "type": "integer"
                          },
                          "totalScore": {
                            "type": "number"
                          },
                          "phoneNumber": {
                            "type": "string",
                            "nullable": true
                          },
                          "universityRole": {
                            "type": "string",
                            "nullable": true
                          },
                          "age": {
                            "type": "integer",
                            "nullable": true
                          },
                          "gender": {
                            "type": "string",
                            "nullable": true
                          },
                          "category": {
                            "type": "string",
                            "nullable": true
                          },
                          "institution": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}/suspend": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Suspends a user account (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/users/{id}/activate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Activates a suspended user account (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "description": "Deletes a user account (Admin only)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/users/provision": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Provisions a new user account (Admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "role"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/users/{id}/approve": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Approves a pending student (Instructor or Admin)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ]
}